+++ for_CRAN/datasets_doc.R
@@ -229,7 +229,7 @@
 #'  This is NOT real biological/ecological data. They were designed for illustrative purposes only.
 #'  The phylogeny used is also NOT a properly time-calibrated phylogeny. It uses an ill-designed old calibration for illustrative purposes.
 #'
-#' @usage Ponerinae_cont_data_old_calib <- readRDS(system.file("extdata", "Ponerinae_cont_data_old_calib", package = "deepSTRAPP"))
+#' @usage Ponerinae_cont_data_old_calib <- readRDS(system.file("extdata", "Ponerinae_cont_data_old_calib.rds", package = "deepSTRAPP"))
 #' @format A list with 5 elements.
 #'
 #' @details A list of five objects containing information on the evolution of fake size data in Ponerinae ants.
@@ -248,7 +248,7 @@
 #'
 #' @seealso [deepSTRAPP::prepare_trait_data()]
 #'
-"Ponerinae_cont_data_old_calib"
+NULL
 
 ### 9/ Categorical trait evolution data for Ponerinae ants using 2-level factor ####
 
@@ -733,7 +733,7 @@
 #'   It summaries STRAPP test results and can optionally include trait and diversification rate data frames, STRAPP test outputs,
 #'   updated trait data objects, and updated BAMM objects, that are required fro downstream plots and analyses.
 #'
-#' @usage Ponerinae_deepSTRAPP_cont_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_cont_old_calib_0_40", package = "deepSTRAPP"))
+#' @usage Ponerinae_deepSTRAPP_cont_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_cont_old_calib_0_40.rds", package = "deepSTRAPP"))
 #' @format A list with 10 elements.
 #'
 #' @details deepSTRAPP output summarizing results of a deepSTRAPP run carried out with [deepSTRAPP::run_deepSTRAPP_over_time()].
@@ -772,7 +772,7 @@
 #' @keywords datasets
 #' @name Ponerinae_deepSTRAPP_cont_old_calib_0_40
 #'
-"Ponerinae_deepSTRAPP_cont_old_calib_0_40"
+NULL
 
 ### 19/ Output from run_deepSTRAPP_over_time on categorical trait with 2-levels ####
 
@@ -783,7 +783,7 @@
 #'   It summaries STRAPP test results and can optionally include trait and diversification rate data frames, STRAPP test outputs,
 #'   updated trait data objects, and updated BAMM objects, that are required fro downstream plots and analyses.
 #'
-#' @usage Ponerinae_deepSTRAPP_cat_2lvl_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_cat_2lvl_old_calib_0_40", package = "deepSTRAPP"))
+#' @usage Ponerinae_deepSTRAPP_cat_2lvl_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_cat_2lvl_old_calib_0_40.rds", package = "deepSTRAPP"))
 #' @format A list with 10 elements.
 #'
 #' @details deepSTRAPP output summarizing results of a deepSTRAPP run carried out with [deepSTRAPP::run_deepSTRAPP_over_time()].
@@ -827,7 +827,7 @@
 #' @keywords datasets
 #' @name Ponerinae_deepSTRAPP_cat_2lvl_old_calib_0_40
 #'
-"Ponerinae_deepSTRAPP_cat_2lvl_old_calib_0_40"
+NULL
 
 ### 20/ Output from run_deepSTRAPP_over_time on categorical trait with 3-levels ####
 
@@ -838,7 +838,7 @@
 #'   It summaries STRAPP test results and can optionally include trait and diversification rate data frames, STRAPP test outputs,
 #'   updated trait data objects, and updated BAMM objects, that are required fro downstream plots and analyses.
 #'
-#' @usage Ponerinae_deepSTRAPP_cat_3lvl_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_cat_3lvl_old_calib_0_40", package = "deepSTRAPP"))
+#' @usage Ponerinae_deepSTRAPP_cat_3lvl_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_cat_3lvl_old_calib_0_40.rds", package = "deepSTRAPP"))
 #' @format A list with 11 elements.
 #'
 #' @details deepSTRAPP output summarizing results of a deepSTRAPP run carried out with [deepSTRAPP::run_deepSTRAPP_over_time()].
@@ -884,7 +884,7 @@
 #' @keywords datasets
 #' @name Ponerinae_deepSTRAPP_cat_3lvl_old_calib_0_40
 #'
-"Ponerinae_deepSTRAPP_cat_3lvl_old_calib_0_40"
+NULL
 
 
 ### 21/ Output from run_deepSTRAPP_over_time on biogeographic ranges ####
@@ -896,7 +896,7 @@
 #'   It summaries STRAPP test results and can optionally include trait and diversification rate data frames, STRAPP test outputs,
 #'   updated trait data objects, and updated BAMM objects, that are required fro downstream plots and analyses.
 #'
-#' @usage Ponerinae_deepSTRAPP_biogeo_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_biogeo_old_calib_0_40", package = "deepSTRAPP"))
+#' @usage Ponerinae_deepSTRAPP_biogeo_old_calib_0_40 <- readRDS(system.file("extdata", "Ponerinae_deepSTRAPP_biogeo_old_calib_0_40.rds", package = "deepSTRAPP"))
 #' @format A list with 10 elements.
 #'
 #' @details deepSTRAPP output summarizing results of a deepSTRAPP run carried out with [deepSTRAPP::run_deepSTRAPP_over_time()].
@@ -938,7 +938,7 @@
 #' @keywords datasets
 #' @name Ponerinae_deepSTRAPP_biogeo_old_calib_0_40
 #'
-"Ponerinae_deepSTRAPP_biogeo_old_calib_0_40"
+NULL