+++ NEWS.md
@@ -1,3 +1,7 @@
+# deepSTRAPP 1.0.0.9000 (development version)
+
+* Current version used to develop new features
+
 # deepSTRAPP 1.0.0
 
 * First release on CRAN